requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
langchain
faiss-cpu
langchain_google_genai #with the help of langchain we will be able to acess api of google-gemini
